# api/swagger.yml
swagger: "2.0"
info:
  version: v1
  title: Birzzha
host: localhost:8000
basePath: /v1
schemes:
  - http
...
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  TokenHeader:
    type: apiKey
    in: header
    name: X-Token
  TokenQuery:
    type: apiKey
    in: query
    name: token
security:
  - TokenHeader: []
  - TokenQuery: []
...
paths:
...
  /reviews:
    get:
      summary: Get Reviews
      operationId: getReviews
      description: Получить список отзывов.
      tags: [landing]
      security: []
      parameters:
        - in: query
          name: offset
          type: integer
          required: false
          description: Офсет отзывов.
        - in: query
          name: limit
          type: integer
          required: false
          description: Лимит отзывов.
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ReviewList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /admin/reviews:
    get:
      summary: Get Reviews
      operationId: adminGetReviews
      description: Получить список отзывов.
      tags: [admin]
      parameters:
        - in: query
          name: offset
          type: integer
          required: false
          description: Офсет отзывов.
        - in: query
          name: limit
          type: integer
          required: false
          description: Лимит отзывов.
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ReviewList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/reviews/{id}:
    parameters:
      - in: path
        name: id
        type: integer
        required: true
        description: ID отзыва

    delete:
      summary: Delete Review
      operationId: adminDeleteReview
      description: Удалить отзыв.
      tags: [admin]
      responses:
        204:
          description: No content
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
    put:
      summary: Update Review
      operationId: adminUpdateReview
      description: Обновить отзыв.
      tags: [admin]

      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/ReviewUpdate"
          required: true
          description: Отредактированный отзыв.

      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Review"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /lots:
    get:
      summary: Get Lots
      operationId: getLots
      tags: [catalog]
      description: Получение списка доступных лотов.
      parameters:
        - in: query
          name: topics
          description: Фильтрация по ID категорий
          type: array
          items:
            type: integer

        - in: query
          name: members_count_from
          description: От кол-ва пдп
          type: integer

        - in: query
          name: members_count_to
          description: До кол-ва пдп
          type: integer

        - in: query
          name: price_from
          description: Цена от
          type: integer

        - in: query
          name: price_to
          description: Цена до
          type: integer

        - in: query
          name: price_per_member_from
          description: Цена пдп от
          type: number

        - in: query
          name: price_per_member_to
          description: Цена пдп до
          type: number

        - in: query
          name: daily_coverage_from
          description: суточный охват от
          type: integer

        - in: query
          name: daily_coverage_to
          description: суточный охват до
          type: integer

        - in: query
          name: price_per_view_from
          description: Цена просмотра от
          type: number

        - in: query
          name: price_per_view_to
          description: Цена просмотра до
          type: number

        - in: query
          name: monthly_income_from
          description: Месячный доход от
          type: integer

        - in: query
          name: monthly_income_to
          description: Месячный доход до
          type: integer

        - in: query
          name: payback_period_from
          description: Окупаемость в месацах от
          type: number

        - in: query
          name: payback_period_to
          description: Окупаемость в месацах до
          type: number

        - in: query
          name: sort_by
          type: string
          enum:
            - members_count
            - price
            - price_per_member
            - price_per_view
            - daily_coverage
            - monthly_income
            - payback_period
            - created_at
          default: created_at

        - in: query
          name: sort_by_type
          type: string
          enum:
            - asc
            - desc
          default: desc
          description: сортировать от меньшего к большему (`asc`), или от большего к меньшему (`desc`)

        - in: query
          name: offset
          type: integer
          description: сдвиг начала каждой страницы на указанное количество единиц

        - in: query
          name: limit
          type: integer
          description: максимальное количество объектов, которое может вернуть запрос

      security:
        - {}
        - TokenHeader: []
        - TokenQuery: []
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /admin/users:
    get:
      summary: Get Users
      operationId: adminGetUsers
      description: Получить список пользователей.
      tags: [admin]
      parameters:
        - in: query
          name: offset
          type: integer
          required: false
          description: Офсет пользователей.
        - in: query
          name: limit
          type: integer
          required: false
          description: Лимит пользователей.
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminFullUserList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/users/{id}/admin:
    parameters:
      - name: id
        in: path
        type: integer
        required: true
        description: ID пользователя
    post:
      summary: Change Admin Rights
      description: Изменение админ прав пользователя.
      operationId: toggleUserAdmin
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminFullUser"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/lot/statuses:
    parameters:
      - name: user_id
        in: query
        type: integer
        required: false
        description: ID пользователя.
    get:
      summary: Get Lot Statuses
      description: |
        Получить статусы лотов и количество лотов по ним.
        В случае если передать user_id, в счетчиках будет кол-во лотов по статусу для конкретного юзера.

      operationId: adminGetLotStatuses
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotStatusesCount"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/lots:
    parameters:
      - name: status
        in: query
        type: string
        required: false
        description: Статус лота.
        enum:
          - created
          - paid
          - published
          - declined
          - canceled
      - name: limit
        in: query
        type: integer
        required: false
        description: Лимит лотов.
      - name: offset
        in: query
        type: integer
        required: false
        description: Офсет лотов.
      - name: user_id
        in: query
        type: integer
        required: false
        description: ID пользователя.

    get:
      summary: Get Lots
      description: Получить лоты.
      operationId: adminGetLots
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminFullLot"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings:
    get:
      summary: Get Settings
      description: Получение настроек.
      operationId: adminGetSettings
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminSettings"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/prices:
    put:
      parameters:
        - in: body
          required: true
          name: prices
          description: Настройки цен.
          schema:
            $ref: "#/definitions/AdminSettingsPrices"

      summary: Update Settings Prices
      description: Обновление настройки цен.
      operationId: adminUpdateSettingsPrices
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminSettingsPrices"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/channel:
    put:
      parameters:
        - in: body
          required: true
          name: channel
          description: Настройки канала.
          schema:
            $ref: "#/definitions/AdminSettingsChannel"

      summary: Update Settings Channel
      description: Обновление настройки канала.
      operationId: adminUpdateSettingsChannel
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminSettingsChannel"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/landing:
    put:
      parameters:
        - in: body
          required: true
          name: landing
          schema:
            $ref: "#/definitions/AdminSettingsLandingInput"
      summary: Update Landing Stats
      operationId: adminUpdateSettingsLanding
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminSettingsLanding"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/topics:
    post:
      parameters:
        - in: body
          required: true
          name: topic
          description: Категория.
          schema:
            $ref: "#/definitions/InputTopic"

      summary: Create Topic
      description: Создание категории.
      operationId: adminCreateTopic
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminTopic"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/topics/{id}:
    parameters:
      - in: path
        required: true
        name: id
        type: integer
        description: ID категории.

    put:
      parameters:
        - in: body
          required: true
          name: topic
          description: Категория.
          schema:
            $ref: "#/definitions/InputTopic"

      summary: Update Topic
      description: Обновление категории.
      operationId: adminUpdateTopic
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/AdminTopic"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"

    delete:
      summary: Delete Topic
      description: Удаление категории.
      operationId: adminDeleteTopic
      tags: [admin]
      responses:
        204:
          description: No content
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/lot-canceled-reasons:
    post:
      parameters:
        - in: body
          required: true
          name: reason
          description: Причина отмены лота.
          schema:
            $ref: "#/definitions/InputLotCanceledReason"

      summary: Create Lot Canceled Reason
      description: Создание причины отмены лота.
      operationId: adminCreateLotCanceledReason
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotCanceledReason"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /admin/settings/lot-canceled-reason/{id}:
    parameters:
      - in: path
        required: true
        name: id
        type: integer
        description: ID причины отмены лота.

    put:
      parameters:
        - in: body
          required: true
          name: reason
          description: Причина отмены лота.
          schema:
            $ref: "#/definitions/InputLotCanceledReason"

      summary: Update Lot Canceled Reason
      description: Обновление причины отмены лота.
      operationId: adminUpdateLotCanceledReason
      tags: [admin]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotCanceledReason"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"

    delete:
      summary: Delete Lot Canceled Reason
      description: Удаление причины отмены лота.
      operationId: adminDeleteLotCanceledReason
      tags: [admin]
      responses:
        204:
          description: No content
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
definitions:
  AdminSettingsLandingInput:
    type: object
    required:
      - unique_users_per_month_shift
      - avg_site_reach_shift
      - avg_channel_reach_shift
    properties:
      unique_users_per_month_shift:
        type: integer
        description: сдвиг кол-ва уникальных посетителей в месяц.
      avg_site_reach_shift:
        type: integer
        description: сдвиг охвата одного лтоа
      avg_channel_reach_shift:
        type: integer
        description: сдвиг охвата одного поста в Telegram
  AdminSettingsLanding:
    type: object
    required:
      - unique_users_per_month_actual
      - unique_users_per_month_shift
      - avg_site_reach_actual
      - avg_site_reach_shift
      - avg_channel_reach_actual
      - avg_channel_reach_shift
    description: |
      Содержит статистику которая показывается на лендинке.
      Поля вида `*_shift` нужны для того чтобы на старте показатель не были слишком низкие.
    properties:
      unique_users_per_month_actual:
        type: integer
        description: реальное кол-во уникальных посетителей в месяц
      unique_users_per_month_shift:
        type: integer
        description: сдвиг кол-ва уникальных посетителей в месяц.
      avg_site_reach_actual:
        type: integer
        description: реальный охват одного лота
      avg_site_reach_shift:
        type: integer
        description: сдвиг охвата одного лтоа
      avg_channel_reach_actual:
        type: integer
        description: реальный средний охват одного поста в канале
      avg_channel_reach_shift:
        type: integer
        description: сдвиг охвата одного поста в Telegram
...
  LotList:
    type: object
    description: Список лотов
    required:
      - total
      - items
    properties:
      total:
        type: integer
        description: кол-во объектов подходщих под фильтр
      items:
        type: array
        items:
          $ref: "#/definitions/LotListItem"
...
  Error:
    type: object
    required:
      - code
      - description
    properties:
      code:
        type: string
        enum: [test]
        description: Код ошибки
      description:
        type: string
        description: Описание ошибки
...
  Review:
    description: Отзыв.
    required:
      - id
      - user
      - text
      - created_at
    properties:
      id:
        type: integer
        description: ID отзыва.
      user:
        $ref: "#/definitions/ReviewUser"
      text:
        type: string
        description: Текст отзыва.
      created_at:
        type: integer
        format: int64
        description: Дата создания отзыва.
  ReviewList:
    description: Отзывы и их количетсво.
    required:
      - total
      - items
    properties:
      total:
        type: integer
        description: Общее количество отзывов.
      items:
        type: array
        items:
          $ref: "#/definitions/Review"
...
  AdminFullUserList:
    type: object
    required:
      - total
      - items
    properties:
      total:
        type: integer
        description: Общее количетсво пользователей.
      items:
        type: array
        items:
          $ref: "#/definitions/AdminFullUser"
  AdminFullUser:
    type: object
    required:
      - id
      - telegram_id
      - avatar
      - full_name
      - username
      - link
      - lots
      - is_admin
      - joined_from
      - joined_at
      - updated_at
    properties:
      id:
        type: integer
        description: ID пользователя.
      telegram_id:
        type: integer
        description: Telegram ID пользователя.
      avatar:
        type: string
        description: Ссылка на аватарку пользователя.
      full_name:
        type: string
        description: Полное имя пользователя.
      username:
        type: string
        description: Юзернейм пользователя.
      link:
        type: string
        format: url
        description: |
          Ссылка на профиль пользователя в Telegram.
          В случае если нет username, пользователя сначала перенаправит в бота,
          а бот уже даст специальную ссылку которая работает только в Telegram.

      lots:
        type: integer
        description: Количество лотов пользователя.
      is_admin:
        type: boolean
        description: Админ ли пользователь.
      joined_from:
        type: string
        enum: [site, bot]
        description: Метод регистрации пользователя.
      joined_at:
        type: integer
        description: Дата регистрации пользователя.
      updated_at:
        type: integer
        description: Дата обновления данных пользователя.
  ReviewUpdate:
    description: Отзыв для обновления.
    required:
      - text
    properties:
      text:
        type: string
        description: Текст отзыва.
...
  LotStatusesCount:
    type: object
    description: Статусы лотов и количество лотов по ним.
    required:
      - created
      - paid
      - published
      - declined
      - canceled
    properties:
      created:
        type: integer
        description: Количество созданых лотов.
      paid:
        type: integer
        description: Количество оплаченых лотов.
      published:
        type: integer
        description: Количество опубликованых лотов.
      declined:
        type: integer
        description: Количество отвергнутых лотов.
      canceled:
        type: integer
        description: Количество отмененных лотов.
...
  AdminFullLot:
    type: object
    description: Список лотов.
    required:
      - total
      - items
    properties:
      total:
        type: integer
        description: Общее количество лотов в соответствии с фильтрами.
      items:
        type: array
        items:
          $ref: "#/definitions/AdminLot"
...
  InputTopic:
    type: object
    description: Категория лота
    required:
      - name
    properties:
      name:
        type: string
        description: Название категории на русском.
...
  AdminSettingsPrices:
    type: object
    required:
      - application
      - change
      - cashier
    properties:
      application:
        $ref: "#/definitions/Money"
      change:
        $ref: "#/definitions/Money"
      cashier:
        type: string
        description: Юзернейм кассы.
  AdminSettingsChannel:
    type: object
    required:
      - id
      - join_link
      - public_username
    properties:
      id:
        type: integer
        description: ID канала.
      join_link:
        type: string
        description: Ссылка на канал.
      public_username:
        type: string
        description: Юзернейм канала.
...
  AdminSettings:
    type: object
    required:
      - prices
      - channel
      - topics
      - lot_canceled_reasons
      - landing
    properties:
      prices:
        $ref: "#/definitions/AdminSettingsPrices"
      channel:
        $ref: "#/definitions/AdminSettingsChannel"
      topics:
        type: array
        items:
          $ref: "#/definitions/AdminTopic"
      lot_canceled_reasons:
        type: array
        items:
          $ref: "#/definitions/LotCanceledReason"
      landing:
        $ref: "#/definitions/AdminSettingsLanding"
...